Refactor parsing of standard config files into a single function.
The names of config files were present in multiple files. Now, they are all in
one place.

Also, instead of relative paths, the new function uses absolute paths to the
config files. This will allow us to use the framework from other locations than
the root directory of the framework.