Add a unification of called function names into our C parser due to builtins.
We have to unify names of builtins, such as memset(), because they are
sometimes parsed as memset() but sometimes as e.g. __builtin___memset_chk().
This makes our regression tests easier to write and read as we can assume that
e.g. memset() is always memset() and not __builtin___memset_chk().